Import from svn r9668.
The SVN repo uses the svn 'externals' mechanism to pull together code
from various places in the repository into a single 'virtualtrunk'
directory, which is the DiFX code itself. We do not replicate that in
git; rather, we simply flatten everything into a single tree.

A more detailed discussion on the travails of importing from svn->git,
and why we haven't imported the history, is recorded separately.