Rewrite mpifxcorr cmd line parsing to use getopt
The main motivation here is to add the --usegpu option, but at the same time it
should make everything more flexible. The command line arguments haven't
changed (although the getopt parser is more liberal in what it accepts, so
variants like -r=12 will now be understood, and order of arguments is no longer
quite so important).