Preparing for GPU -- surrounding framework
Add a separate process_gpu() function to 'Mode' which gets called if we are
running the GPU version; as well as surrounding infrastructure to get it to
compile in CUDA and non-CUDA environments.

For now, the process_gpu() does the same thing as regular process(), but on the
CPU (i.e. it's a straight copy of the code).

Core::processdata() and associated data structures move from 'private' to
'protected' so they can be override in the GPU base class (and 'execute' for
now is no longer virtual).