FFTW proof of concept
Do the FFT using FFTW, rather than IPP.

This was used to check whether the difference between cuFFT and IPP was
significant.

Note that it is definitely not production-ready (it recreates the plan
in a hot loop, and leaks a lot of memory) - it is a proof of concept
only. You will also manually need to add -lfftw3_threads -lfftw3 -lm to
the Makefile.